requirements.txt
# Core data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning and clustering
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# LLM integration (choose based on your preference)
openai>=1.0.0
google-genai>=0.2.0

# Development and analysis tools
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Data loading and parsing
lxml>=4.9.0
py7zr>=0.20.0

# Progress bars and utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: Advanced topic modeling (uncomment if needed)
# bertopic>=0.15.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0 